[*] added casting of input parameters.
[interpreter.file_system.c] decreased area of visibility of
entry_types_str variable.

[string_unit.c] removed unnecessary assigned of the pos variable.

[xml.c] replaced start variable with pos at the xml_get_tag_finish_pos
function at one of the if case.

[.github/workflows/cmake.yml] refactored build script for CMake work
flow.

+++ date_time.c
@@ -364,12 +364,12 @@ uint8_t datetime_to_string(
 	int inputs[6];
 	inputs[0] = day;
 	inputs[1] = month;
-	inputs[2] = year;
+	inputs[2] = (int)year;
 	inputs[3] = hour;
 	inputs[4] = minute;
 	inputs[5] = second;
 #else
-	int inputs[] = { day, month, year, hour, minute, second };
+	int inputs[] = { day, month, (int)year, hour, minute, second };
 #endif
 	uint8_t* output_ = buffer_data(output, size);
 	return buffer_resize(output, size + datetime_to_char_array(inputs, output_));
